requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.14.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
Faker==17.0.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
validate-docbr==1.10.0
